Check that sensors have power; count sensor errors
DS18B20s can be powered in two ways: with 5v supplied directly to a pin,
or by using power from the data line.

In the "parasite power" mode, the bus master must provide a strong
pullup to +5v during some operations, such as temperature conversions
and accesses to the EEPROM.  Our hardware doesn't support this, so any
device that doesn't have 5v power supplied directly will produce bogus
temperature readings with correct CRCs.

Fortunately there's a command that can check whether a DS18B20 is
powered directly.  We use this before relying on a reading.

We also keep a count of the number of times we find unpowered devices,
or devices that give a bad CRC.  This could be an early indication of
probe assembly failure.  A later commit will add a way to access and
reset these error counts over the RS485 bus.