Allow null readings to be stored in the database
The latest firmware will return the string "None" if a sensor is faulty.
(This is different to the python value None, which is returned when a
controller is missing.)

Store these readings in the database as null values, so that periods of
the sensor being faulty can be shown on the graph.

Requires a database update:
alter table datalog_floatdatum alter column data drop not null;
alter table datalog_integerdatum alter column data drop not null;
alter table datalog_stringdatum alter column data drop not null;